week2

+++ src/main/java/ru/bobrysheva/calcapp/_main/_Main.java
@@ -6,18 +6,15 @@ import ru.bobrysheva.calcapp.util.Calculator;
 import java.util.Scanner;
 
 public class _Main {
-    public static void main(String[] args){
+    public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int a = sc.nextInt();
         int b = sc.nextInt();
-  //      line operationSign = in.nextLine();
-        int result = Calculator.sum(a,b);
+        int result = Calculator.sum(a, b);
         ResultWriterService.printResult(a, b, result, "сложение", "+");
-        double result2 = Calculator.div(a,b);
+        int result2 = Calculator.div(a, b);
         ResultWriterService.printResult(a, b, result2, "деление", "/");
-        int result3 = Calculator.sub(a,b);
+        int result3 = Calculator.sub(a, b);
         ResultWriterService.printResult(a, b, result3, "вычитание", "-");
-
-
     }
 }
